resolver component separated, resolver generated and registered only if requests are present
