+++ oarepo_model_builder_requests/datatypes/components/requests_model/requests.py
@@ -6,35 +6,6 @@ from oarepo_model_builder.utils.camelcase import camel_case, snake_case
 from oarepo_model_builder.validation.utils import ImportSchema
 
 
-class RecordResolverClassSchema(ma.Schema):
-    class Meta:
-        unknown = ma.RAISE
-
-    generate = ma.fields.Bool()
-    class_ = ma.fields.Str(
-        attribute="class",
-        data_key="class",
-    )
-    base_classes = ma.fields.List(
-        ma.fields.Str(),
-        attribute="base-classes",
-        data_key="base-classes",
-    )
-    extra_code = ma.fields.Str(
-        attribute="extra-code",
-        data_key="extra-code",
-    )
-    module = ma.fields.String(metadata={"doc": "Class module"})
-    imports = ma.fields.List(
-        ma.fields.Nested(ImportSchema), metadata={"doc": "List of python imports"}
-    )
-    custom_proxy_class = ma.fields.Str(
-        attribute="custom-proxy-class",
-        data_key="custom-proxy-class",
-    )
-    skip = ma.fields.Boolean()
-
-
 class RequestActionSchema(ma.Schema):
     class Meta:
         unknown = ma.RAISE
@@ -100,45 +71,12 @@ class RequestsComponent(DataTypeComponent):
             attribute="requests",
             data_key="requests",
         )
-        record_resolver = ma.fields.Nested(
-            RecordResolverClassSchema,
-            attribute="record-resolver",
-            data_key="record-resolver",
-        )
 
     def before_model_prepare(self, datatype, *, context, **kwargs):
         module = datatype.definition["module"]["qualified"]
         profile_module = context["profile_module"]
 
         requests = set_default(datatype, "requests", {})
-        # resolver
-        record_resolver = set_default(datatype, "record-resolver", {})
-
-        record_resolver.setdefault("generate", True)
-        resolver_module = record_resolver.setdefault(
-            "module", f"{module}.{profile_module}.requests.resolvers"
-        )
-        record_resolver.setdefault(
-            "class",
-            f"{resolver_module}.{datatype.definition['module']['prefix']}Resolver",
-        )
-        record_resolver.setdefault(
-            "base-classes",
-            ["RecordResolver"],
-        )
-        record_resolver.setdefault(
-            "imports",
-            [
-                {
-                    "import": "invenio_records_resources.references.RecordResolver",
-                },
-            ],
-        )
-
-        if context["profile"] == "draft":
-            record_resolver.setdefault(
-                "custom-proxy-class", "oarepo_runtime.resolvers.DraftProxy"
-            )
 
         for request_name, request_input_data in requests.items():
             request_module = f"{module}.{profile_module}.requests.{snake_case(request_name).replace('-', '_')}"
