+++ oarepo_model_builder_requests/invenio/templates/invenio_requests_config.py.jinja2
@@ -8,8 +8,9 @@ from invenio_users_resources.resolvers import UserResolver
 
 
 REQUESTS_REGISTERED_TYPES = [{% for request_data in vars.requests.values() %}{{ request_data.type.class|base_name }}, {% endfor %}]
-# drafts - make a special variable for resolvers
+{% if not vars.record_resolver.skip %}
 REQUESTS_ENTITY_RESOLVERS = [
         UserResolver(),
         {{ vars.record_resolver.class|base_name }}(record_cls={{ vars.record.class|base_name }}, service_id="{{ vars.service_config.service_id }}", type_key="{{ vars.module.prefix_snake }}")
     ]
+{% endif %}
